Refactored Enemy Spawner
Enemy spawner now contains an array of wavePrefabs rows, which in turn contains the prefabs of all the enemies for that wave, in the order they spawn. This allows us to create waves and add enemies to waves through unity, and makes the wave management more scaleable. Also added a Debug.Log() for when the level is finished, ie. all waves are completed